商品入库

+++ op-system-beans/src/main/java/com/hf/op/infrastructure/dto/department/GoodsShelvesGoodsRqDto.java
@@ -31,7 +31,7 @@ public class GoodsShelvesGoodsRqDto  implements Serializable {
 	/**
 	 * 数量
 	 */
-	private BigDecimal num;
+	private Integer num;
 
 	/**
 	 * 上架价格
